Fixed Link not working in footer
Removed Footer component from App, and added it to Home and PrivacyPolicy.
I did this because the Navbar component is located in each page, not in the App component. So it makes sense to put the Footer component in the same place.

PS:
Link in README.md went to a 404 page. Removed 's' from the link to fix.